requirements.txt
langchain==1.2.9
langchain-community==0.4.1
langsmith==0.7.1
langchain-deepseek==1.0.1
mysql-connector-python==9.6.0
openpyxl==3.1.5
pandas==3.0.1
SQLAlchemy==2.0.47
Flask==3.1.3
Flask-Limiter==4.1.1
python-dotenv==1.2.1
gunicorn==25.1.0